complete the kl algorithm!

+++ algorithms_matlab/include/mystd.h
@@ -129,7 +129,7 @@ int edge_matrix_double(double **adj,int row,int column);
 int min_double(double *D,double *min,int length);
 double trace(double **adj,int row);
 int *unique(int *adj,int length,int *r_length);
-int **inv(int **adj,int row,int column);
+double **inv(int **adj,int row,int column);
 int find_num(int *adj,int length,int num);
 #endif
